pyproject.toml
[tool.poetry]
name = "flask-project-demo"
version = "0.1.0"
description = "A demo flask project for teaching purposes"
authors = ["Matteo Spanio <spanio@dei.unipd.it>"]
license = "MIT"
readme = "README.md"

[tool.poetry.scripts]
server-cli = "flask_project_demo.cli:cli"

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^3.0.2"
mkdocs = "^1.5.3"
sqlalchemy = "^2.0.29"
python-dotenv = "^1.0.1"
faker = "^24.4.0"
marshmallow = "^3.21.1"
flask-jwt-extended = "^4.6.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"
pytest = "^8.1.1"
ruff = "^0.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
